Extract RackGameView + NonGolfGameView from their routes (Spec 2 Phase 2)
Apply the #545 recipe to rack + non-golf: move each game view into a re-hostable
component (reads its own tripId + ?game=), route pages become thin wrappers. Pure
refactor — no behavior change. Same component now serves the route AND the
leaderboard panel (wired next commit).
